backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
slowapi
scikit-learn
xgboost
openai
groq
passlib
python-jose
pandas
numpy
joblib 